+++ lib/models/asyncapi.js
@@ -9,7 +9,8 @@ const Components = require('./components');
 const MixinExternalDocs = require('../mixins/external-docs');
 const MixinTags = require('../mixins/tags');
 const MixinSpecificationExtensions = require('../mixins/specification-extensions');
-const {xParserMessageName, xParserSchemaId, xParserCircle, xParserCircleProps} = require('../constants');
+const {xParserCircle, xParserCircleProps} = require('../constants');
+const {assignNameToAnonymousMessages, assignNameToComponentMessages, assignUidToComponentSchemas, assignUidToParameterSchemas, assignIdToAnonymousSchemas} = require('../anonymousNaming');
 const {traverseAsyncApiDocument, SchemaIteratorCallbackType} = require('../iterators');
 
 /**
@@ -267,99 +268,6 @@ function markCircularSchemas(doc) {
     }
   };
   traverseAsyncApiDocument(doc, circularCheckCallback);
-}
-
-/**
- * Assign message keys as message name to all the component messages.
- * 
- * @private
- * @param {AsyncAPIDocument} doc 
- */
-function assignNameToComponentMessages(doc) {
-  if (doc.hasComponents()) {
-    for (const [key, m] of Object.entries(doc.components().messages())) {
-      if (m.name() === undefined) {
-        m.json()[String(xParserMessageName)] = key;
-      }
-    }
-  }
-}
-
-/**
- * Assign parameter keys as uid for the parameter schema.
- * 
- * @private
- * @param {AsyncAPIDocument} doc 
- */
-function assignUidToParameterSchemas(doc) {
-  doc.channelNames().forEach(channelName => {
-    const channel = doc.channel(channelName);
-    for (const [parameterKey, parameterSchema] of Object.entries(channel.parameters())) {
-      parameterSchema.json()[String(xParserSchemaId)] = parameterKey;
-    }
-  });
-}
-
-/**
- * Assign uid to component schemas. 
- * 
- * @private
- * @param {AsyncAPIDocument} doc 
- */
-function assignUidToComponentSchemas(doc) {
-  if (doc.hasComponents()) {
-    for (const [key, s] of Object.entries(doc.components().schemas())) {
-      s.json()[String(xParserSchemaId)] = key;
-    }
-  }
-}
-
-/**
- * Assign anonymous names to nameless messages.
- * 
- * @private
- * @param {AsyncAPIDocument} doc 
- */
-function assignNameToAnonymousMessages(doc) {
-  let anonymousMessageCounter = 0;
-
-  if (doc.hasChannels()) {
-    doc.channelNames().forEach(channelName => {
-      const channel = doc.channel(channelName);
-      if (channel.hasPublish()) addNameToKey(channel.publish().messages(), ++anonymousMessageCounter);
-      if (channel.hasSubscribe()) addNameToKey(channel.subscribe().messages(), ++anonymousMessageCounter);
-    });
-  }
-}
-
-/**
- * Add anonymous name to key if no name provided.
- * 
- * @private
- * @param {Message} map of messages 
- */
-function addNameToKey(messages, number) {
-  messages.forEach(m => {
-    if (m.name() === undefined) {
-      m.json()[String(xParserMessageName)] = `<anonymous-message-${number}>`;
-    }
-  });
-}
-
-/**
- * Gives schemas id to all anonymous schemas.
- * 
- * @private
- * @param {AsyncAPIDocument} doc 
- */
-function assignIdToAnonymousSchemas(doc) {
-  let anonymousSchemaCounter = 0;
-  const callback = (schema) => {
-    if (!schema.uid()) {
-      schema.json()[String(xParserSchemaId)] = `<anonymous-schema-${++anonymousSchemaCounter}>`;
-    }
-  };
-  traverseAsyncApiDocument(doc, callback);
 }
 
 module.exports = mix(AsyncAPIDocument, MixinTags, MixinExternalDocs, MixinSpecificationExtensions);
